chore(folder-structure): reorg decisions and put json files in a common dir
