--- decision-ia/llm-playground/project-description-sample.json
@@ -1,10 +1,0 @@
-{
-  "description_projet_01": "Revitalisation d'une zone humide",
-  "description_projet_02": "Entretient d'un vieux moulin",
-  "description_projet_03": "Réhabilitation d'une ancienne école en lieu dédié à la santé.",
-  "description_projet_04": "Voir fiche action PVD /ORT n°21 Après une première phase d’aménagement, la commune poursuit son action dans le cadre d’une 2ème phase d’aménagement. Cette prochaine étape concerne : - La mise en accessibilité du site aux personnes en situation de handicap et l’installation d’équipements en accès libre favorisant la pratique physique. - La protection de l’environnement par l’installation de sanitaires autonomes et la création d’un site O déchets - La création d'un parcours pédagogique",
-  "description_projet_05": "La commune dispose d'outils numériques qu'il est nécessaire d'optimiser et coordonner",
-  "description_projet_06": "Création d’un sentier thématique sur la forêt à Rieutord et sur le patrimoine à Usclades",
-  "description_projet_07": "Je souhaite refaire la voirie communal",
-  "description_projet_08": "Le projet consiste en l’aménagement d’un terrain communal en cœur de commune, jouxtant une structure crèche multi-accueil et un pôle culturel (médiathèque avec une salle attenante). La commune porte le projet d’aménager le reste du site, afin d’offrir aux habitants des espaces de loisirs et de rencontres intergénérationnelles, en lien à la fois avec ces nouvelles structures, mais également avec les usages existants. Avec l’implication des élus du Conseil Municipal des Jeunes, plusieurs hypothèses sont imaginées avec la volonté d’installer / aménager des jeux pour jeunes enfants, des espaces de rencontres, des cheminements, des dispositifs en faveur de la biodiversité, des arbres fruitiers, d’instaurer l’opération « Une naissance, un arbre"
-}
