--- decision-ia/llm-playground/requierment.txt
@@ -1,8 +1,0 @@
-ipykernel
-jupyter
-numpy
-pandas
-plotly
-python-decouple
-markdownify
-aiohttp
