+++ llm-playground/.gitignore
@@ -1,0 +1,12 @@
+.env
+.conda
+hard-database/
+__pycache__/
+generation-score-async.ipynb
+database-studies.ipynb
+rag_prototype.ipynb
+generation-score-proto.ipynb
+project-description-sample/
+graph/
+legacy/
+ai-assistants-proto.json
