+++ llm-playground/report.md
@@ -1,0 +1,118 @@
+---
+
+## Création d’un sentier thématique sur la forêt à Rieutord et sur le patrimoine à Usclades
+
+- **Nom de l'aide** : Valoriser les atouts culturels, touristiques et patrimoniaux du territoire
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/)
+  - **Score**: 80 / 100
+
+  - **Orginal score**: 160
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+---
+
+## Entretient d'un vieux moulin
+
+- **Nom de l'aide** : Valoriser les atouts culturels, touristiques et patrimoniaux du territoire
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/)
+  - **Score**: 70 / 100
+
+  - **Orginal score**: 140
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+---
+
+## Le projet consiste en l’aménagement d’un terrain communal en cœur de commune, jouxtant une structure crèche multi-accueil et un pôle culturel (médiathèque avec une salle attenante). La commune porte le projet d’aménager le reste du site, afin d’offrir aux habitants des espaces de loisirs et de rencontres intergénérationnelles, en lien à la fois avec ces nouvelles structures, mais également avec les usages existants. Avec l’implication des élus du Conseil Municipal des Jeunes, plusieurs hypothèses sont imaginées avec la volonté d’installer / aménager des jeux pour jeunes enfants, des espaces de rencontres, des cheminements, des dispositifs en faveur de la biodiversité, des arbres fruitiers, d’instaurer l’opération « Une naissance, un arbre
+
+- **Nom de l'aide** : Mener des actions en faveur de la cohésion sociale
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/actions-en-faveur-de-la-cohesion-sociale-politique-territoriale-2022-2028/](https://aides-territoires.beta.gouv.fr/aides/actions-en-faveur-de-la-cohesion-sociale-politique-territoriale-2022-2028/)
+  - **Score**: 60 / 100
+
+  - **Orginal score**: 120
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+- **Nom de l'aide** : Accompagner la transition écologique
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables/)
+  - **Score**: 75 / 100
+
+  - **Orginal score**: 150
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+- **Nom de l'aide** : Valoriser les atouts culturels, touristiques et patrimoniaux du territoire
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/)
+  - **Score**: 65 / 100
+
+  - **Orginal score**: 130
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+---
+
+## Revitalisation d'une zone humide
+
+- **Nom de l'aide** : Accompagner la transition écologique
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables/)
+  - **Score**: 80 / 100
+
+  - **Orginal score**: 160
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+---
+
+## Voir fiche action PVD /ORT n°21 Après une première phase d’aménagement, la commune poursuit son action dans le cadre d’une 2ème phase d’aménagement. Cette prochaine étape concerne : - La mise en accessibilité du site aux personnes en situation de handicap et l’installation d’équipements en accès libre favorisant la pratique physique. - La protection de l’environnement par l’installation de sanitaires autonomes et la création d’un site O déchets - La création d'un parcours pédagogique
+
+- **Nom de l'aide** : Mener des actions en faveur de la cohésion sociale
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/actions-en-faveur-de-la-cohesion-sociale-politique-territoriale-2022-2028/](https://aides-territoires.beta.gouv.fr/aides/actions-en-faveur-de-la-cohesion-sociale-politique-territoriale-2022-2028/)
+  - **Score**: 75 / 100
+
+  - **Orginal score**: 150
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+- **Nom de l'aide** : Accompagner la transition écologique
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables/)
+  - **Score**: 65 / 100
+
+  - **Orginal score**: 130
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
+
+- **Nom de l'aide** : Valoriser les atouts culturels, touristiques et patrimoniaux du territoire
+
+  - **URL**: [https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/](https://aides-territoires.beta.gouv.fr/aides/accompagner-le-developpement-dune-production-et-dune-consommation-responsables-1/)
+  - **Score**: 75 / 100
+
+  - **Orginal score**: 150
+
+  - **Error made**: 0
+
+  - **Scoring made**: 2 / 2
