+++ llm-playground/requierment.txt
@@ -1,0 +1,8 @@
+ipykernel
+jupyter
+numpy
+pandas
+plotly
+python-decouple
+markdownify
+aiohttp
